requirements.txt
fastapi==0.95.2
langchain==0.0.220
python-dotenv==0.21.0
uvicorn==0.18.3
openai